refactor: use long for phrase count

+++ src/main/java/com/mirocoder/habittracker/DailyPhraseController.java
@@ -83,7 +83,7 @@ public class DailyPhraseController {
     }
 
     @GetMapping("/daily-phrases/count")
-    public int getPhrasesAmount(){
+    public long getPhrasesAmount(){
         return dailyPhraseService.getPhrasesAmount();
     }
