+++ src/main/java/com/mirocoder/habittracker/service/DailyPhraseService.java
@@ -58,7 +58,7 @@ public class DailyPhraseService {
         return dailyPhraseRepository.findPhraseForToday();
     }
 
-    public int getPhrasesAmount() {
+    public long getPhrasesAmount() {
         return dailyPhraseRepository.countDailyPhrases();
     }
 }
